+++ components/clarinet-deployments/Cargo.toml
@@ -8,7 +8,7 @@ serde = "1"
 serde_json = "1"
 serde_derive = "1"
 clarity_repl = { package = "clarity-repl", path = "../clarity-repl", default-features = false }
-clarinet_files =  { package = "clarinet-files", path = "../clarinet-files", default-features = false }
+clarinet_files = { package = "clarinet-files", path = "../clarinet-files", default-features = false }
 orchestra_types = { package = "orchestra-types", path = "../../vendor/orchestra-types-rs" }
 serde_yaml = "0.8.23"
