+++ components/clarinet-files/Cargo.toml
@@ -13,10 +13,9 @@ bip39 = { version = "1.0.1", default-features = false }
 clarity_repl = { package = "clarity-repl", path = "../clarity-repl", default-features = false }
 libsecp256k1 = "0.7.0"
 toml = { version = "0.5.6", features = ["preserve_order"] }
-url =  { version = "2.2.2", features = ["serde"] }
+url = { version = "2.2.2", features = ["serde"] }
 tiny-hderive = "0.3.0"
 bitcoin = { version = "0.28.1", optional = true }
-async-trait = "0.1"
 
 [features]
 default = ["cli"]
