+++ components/clarity-lsp/Cargo.toml
@@ -1,36 +1,72 @@
 [package]
+edition = "2021"
 name = "clarity-lsp"
 version = "0.1.0"
-edition = "2021"
 
 [dependencies]
+clarinet_deployments = { package = "clarinet-deployments", path = "../clarinet-deployments", default-features = false }
+clarinet_files = { package = "clarinet-files", path = "../clarinet-files", default-features = false }
+clarity_repl = { package = "clarity-repl", path = "../clarity-repl", default-features = false }
 lsp-types = "0.93"
 orchestra_types = { package = "orchestra-types", path = "../../vendor/orchestra-types-rs" }
-clarinet_files =  { package = "clarinet-files", path = "../clarinet-files", default-features = false }
-clarity_repl = { package = "clarity-repl", path = "../clarity-repl", default-features = false }
-clarinet_deployments = { package = "clarinet-deployments", path = "../clarinet-deployments", default-features = false }
+serde = { version = "1" }
 
 # WASM
-wasm-bindgen = { version = "0.2", features = ["serde-serialize"] }
-wasm-bindgen-futures = { version  = "0.4" }
-serde = "1"
-serde_json = "1"
-serde-wasm-bindgen = "0.4"
-js-sys = "0.3"
-web-sys = { version = "0.3", features = ["console"] }
-futures = "0.3"
-async-trait = "0.1"
-console_error_panic_hook = "0.1.7"
+console_error_panic_hook = { version = "0.1", optional = true }
+js-sys = { version = "0.3", optional = true }
+serde-wasm-bindgen = { version = "0.4", optional = true }
+wasm-bindgen = { version = "0.2", optional = true }
+wasm-bindgen-futures = { version = "0.4", optional = true }
+web-sys = { version = "0.3", features = ["console"], optional = true }
 
 [features]
 default = ["cli"]
-cli = ["clarinet_deployments/cli", "clarinet_deployments/cli", "clarity_repl/cli", "clarinet_files/cli"]
-wasm = ["clarinet_deployments/wasm", "clarity_repl/wasm", "clarinet_files/wasm"]
+cli = [
+  "clarinet_deployments/cli",
+  "clarinet_deployments/cli",
+  "clarity_repl/cli",
+  "clarinet_files/cli",
+]
+wasm = [
+  "clarinet_deployments/wasm",
+  "clarity_repl/wasm",
+  "clarinet_files/wasm",
+  "wasm-bindgen",
+  "wasm-bindgen-futures",
+  "serde-wasm-bindgen",
+  "js-sys",
+  "web-sys",
+  "console_error_panic_hook",
+]
 
 [lib]
+crate-type = ["cdylib", "rlib"]
 name = "clarity_lsp"
 path = "src/lib.rs"
-crate-type = ["cdylib", "rlib"]
+
+[package.metadata.wasm-pack.profile.dev]
+wasm-opt = false
+
+[package.metadata.wasm-pack.profile.dev.wasm-bindgen]
+debug-js-glue = false
+demangle-name-section = true
+dwarf-debug-info = false
+
+[package.metadata.wasm-pack.profile.profiling]
+wasm-opt = ['-O']
+
+[package.metadata.wasm-pack.profile.profiling.wasm-bindgen]
+debug-js-glue = false
+demangle-name-section = true
+dwarf-debug-info = false
+
+[package.metadata.wasm-pack.profile.release]
+# Configuration can be set to an array of strings which are explicit arguments
+# to pass to `wasm-opt`. For example `['-Os']` would optimize for size
+# while `['-O4']` would execute very expensive optimizations passes
+# wasm-opt = ['-Os']
+# TODO(hugocaillard) - benchmark optimizations options
+wasm-opt = false
 
 [package.metadata.wasm-pack.profile.release.wasm-bindgen]
 debug-js-glue = false
