+++ components/clarity-repl/Cargo.toml
@@ -2,19 +2,23 @@
 name = "clarity-repl"
 version = "0.33.0"
 description = "Clarity REPL"
-authors = ["Ludo Galabru <ludo@hiro.so>", "Brice Dobry <brice@hiro.so>", "Reed Rosenbluth <reed@hiro.so>"]
+authors = [
+    "Ludo Galabru <ludo@hiro.so>",
+    "Brice Dobry <brice@hiro.so>",
+    "Reed Rosenbluth <reed@hiro.so>",
+]
 readme = "README.md"
 edition = "2018"
 license = "GPL-3.0-only"
 keywords = ["blockstack", "blockchain", "clarity", "smart-contract", "repl"]
-exclude = [
-    "vs-client/**",
-    ".husky",
-    ".git*"
-    ]
+exclude = ["vs-client/**", ".husky", ".git*"]
 homepage = "https://hirosystems.github.io/clarity-repl/"
 repository = "https://github.com/hirosystems/clarity-repl"
-categories = ["command-line-utilities", "development-tools", "development-tools::build-utils"]
+categories = [
+    "command-line-utilities",
+    "development-tools",
+    "development-tools::build-utils",
+]
 
 [dependencies]
 ansi_term = "0.12.1"
@@ -59,7 +63,6 @@ name = "clarity_repl"
 path = "src/lib.rs"
 # Default type
 # crate-type = ["lib"]
-
 # Use this instead for WASM builds
 crate-type = ["cdylib", "rlib"]
 
@@ -84,7 +87,5 @@ wasm = ["wasm-bindgen", "wasm-bindgen-futures"]
 
 [package.metadata.wasm-pack.profile.release.wasm-bindgen]
 debug-js-glue = false
-demangle-name-section = true
+demangle-name-section = false
 dwarf-debug-info = false
-debug = true
-
